# codigo_laravel-master/codigo_laravel-master/Imobillus/imobillus-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Horse4U - Mais Code - API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/singin": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login bearer",
                "description": "Rota de login de usuário por meio de login e senha",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "suporte@maiscode.com.br",
                                    "password": "123456789"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returna Token de acesso do usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "{{TOKEN}}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "noauth": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}